updated highcharts visulizations
